chore: refactor so we don't keep all feature flags in local
