+++ src/core/client/framework/lib/relay/local.graphql
@@ -1,15 +1,10 @@
 # This local.graphql extension is used by all bundles.
 # These values are all initialized by calling `initLocalBaseState()`
 
-type StaticConfig {
-  featureFlags: FEATURE_FLAG!
-}
-
 type Local {
   accessToken: String
   accessTokenExp: Int
   accessTokenJTI: String
-  staticConfig: StaticConfig!
 }
 
 extend type Query {
