+++ src/core/client/stream/tabs/Comments/helpers/useStaticFlattenReplies.ts
@@ -1,19 +1,14 @@
 import { graphql } from "react-relay";
 
 import { useLocal } from "coral-framework/lib/relay";
-import { GQLFEATURE_FLAG } from "coral-framework/schema";
 
 import { useStaticFlattenReplies } from "coral-stream/__generated__/useStaticFlattenReplies.graphql";
 
 export default function useStaticFlattenReplies() {
   const [local] = useLocal<useStaticFlattenReplies>(graphql`
     fragment useStaticFlattenReplies on Local {
-      staticConfig {
-        featureFlags
-      }
+      flattenReplies
     }
   `);
-  return local.staticConfig.featureFlags.includes(
-    GQLFEATURE_FLAG.FLATTEN_REPLIES
-  );
+  return local.flattenReplies;
 }
